initial backend development

+++ Controllers/UsersController.cs
@@ -10,6 +10,7 @@ using Microsoft.EntityFrameworkCore;
 using RestaurantSystem.Data;
 using RestaurantSystem.Models;
 using RestaurantSystem.Models.Repositories;
+using RestaurantSystem.Services;
 
 namespace RestaurantSystem.Controllers
 {
@@ -18,14 +19,17 @@ namespace RestaurantSystem.Controllers
     public class UsersController : ControllerBase
     {
         private readonly UserRepository userRepository;
+        private readonly PermissionValidation _permissionValidation;
 
         public UsersController(RestaurantSystemContext context, UserManager<User> userManager)
         {
             this.userRepository = new UserRepository(context, userManager);
+            _permissionValidation = new PermissionValidation(context, userManager);
         }
 
         // GET: api/Users
         [HttpGet]
+        [Authorize(Roles = "RestaurantManager")]
         public async Task<ActionResult<IEnumerable<User>>> GetUsersAsync()
         {
             var users = await this.userRepository.GetAllAsync();
@@ -33,8 +37,8 @@ namespace RestaurantSystem.Controllers
         }
 
         // GET: api/Users/5
-        [Authorize]
         [HttpGet("{id}")]
+        [Authorize(Roles = "RestaurantManager, RestaurantEveryDayUse")]
         public async Task<ActionResult<User>> GetUserAsync(string id)
         {
             var user = await this.userRepository.GetAsync(id);
@@ -50,13 +54,20 @@ namespace RestaurantSystem.Controllers
         // PUT: api/Users/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
+        [Authorize(Roles = "Customer, RestaurantManager, RestaurantEveryDayUse")]
         public async Task<ActionResult> PutUserAsync(string id, Register register)
         {
             if (id != register.SystemId)
             {
                 return BadRequest();
             }
-            
+
+            var currentUserEmail = HttpContext.User.Claims.FirstOrDefault(claim => claim.Type == System.Security.Claims.ClaimTypes.Email).Value;
+            if (! await _permissionValidation.isUserTheSameAsync(id, currentUserEmail))
+            {
+                return Unauthorized();
+            }
+
             var returnedUser = await this.userRepository.UpdateAsync(register);
             
             if (returnedUser == null)
@@ -69,8 +80,15 @@ namespace RestaurantSystem.Controllers
 
         // DELETE: api/Users/5
         [HttpDelete("{id}")]
+        [Authorize(Roles = "Customer")]
         public async Task<IActionResult> DeleteUser(string id)
         {
+            var currentUserEmail = HttpContext.User.Claims.FirstOrDefault(claim => claim.Type == System.Security.Claims.ClaimTypes.Email).Value;
+            if (!await _permissionValidation.isUserTheSameAsync(id, currentUserEmail))
+            {
+                return Unauthorized();
+            }
+
             var returnedUser = await this.userRepository.DeleteAsync(id);
             
             if (returnedUser == null)
